# docs/swagger-api-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Banking System API",
    "description": "Sebuah sistem bank sederhana yang mengimplementasikan hal-hal yang sudah dipelajari dari pembelajaran Back End JavaScript - Binar Academy.",
    "contact": {
      "email": "ahmadsubhan@apps.ipb.ac.id"
    },
    "version": "v1"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Transfer between user accounts",
        "description": "",
        "operationId": "createTransaction",
        "requestBody": {
          "description": "Transfer between user accounts by account id",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "integer",
                    "example": 1
                  },
                  "destinationAccountId": {
                    "type": "integer",
                    "example": 2
                  },
                  "amount": {
                    "type": "integer",
                    "example": 5000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transaction successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "amount": {
                          "type": "integer",
                          "example": 5000
                        },
                        "sourceAccountId": {
                          "type": "integer",
                          "example": 1
                        },
                        "destinationAccountId": {
                          "type": "integer",
                          "example": 2
                        },
                        "sourceAccount": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "bankName": {
                              "type": "string",
                              "example": "BRI"
                            },
                            "bankAccountNumber": {
                              "type": "string",
                              "example": "5644386582"
                            },
                            "balance": {
                              "type": "integer",
                              "example": 495000
                            },
                            "userId": {
                              "type": "integer",
                              "example": 1
                            }
                          }
                        },
                        "destinationAccount": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 2
                            },
                            "bankName": {
                              "type": "string",
                              "example": "BRI"
                            },
                            "bankAccountNumber": {
                              "type": "string",
                              "example": "7864539765"
                            },
                            "balance": {
                              "type": "integer",
                              "example": 505000
                            },
                            "userId": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Transaction failed"
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get all transactions",
        "description": "Get list of all transactions.",
        "operationId": "getAllTransactions",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All transactions retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "amount": {
                            "type": "integer",
                            "example": 5000
                          },
                          "sourceAccountId": {
                            "type": "integer",
                            "example": 1
                          },
                          "destinationAccountId": {
                            "type": "integer",
                            "example": 2
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieved all transactions"
          }
        }
      }
    },
    "/transactions/{transaction}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get transaction by transaction id",
        "description": "",
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "transaction",
            "in": "path",
            "description": "The id that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transaction retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "amount": {
                          "type": "integer",
                          "example": 5000
                        },
                        "sourceAccountId": {
                          "type": "integer",
                          "example": 1
                        },
                        "destinationAccountId": {
                          "type": "integer",
                          "example": 2
                        },
                        "sourceAccount": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "bankName": {
                              "type": "string",
                              "example": "BRI"
                            },
                            "bankAccountNumber": {
                              "type": "string",
                              "example": "5644386582"
                            },
                            "balance": {
                              "type": "integer",
                              "example": 495000
                            },
                            "userId": {
                              "type": "integer",
                              "example": 1
                            }
                          }
                        },
                        "destinationAccount": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 2
                            },
                            "bankName": {
                              "type": "string",
                              "example": "BRI"
                            },
                            "bankAccountNumber": {
                              "type": "string",
                              "example": "7864539765"
                            },
                            "balance": {
                              "type": "integer",
                              "example": 505000
                            },
                            "userId": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrived transaction"
          }
        }
      }
    }
  }
}